refactor: split overture, province overlay, and labour economy parts (#3878)
Extract phase-two overture UI, province tab intel/unrevealed assembly, and
production labour projection/mutation helpers into library parts; refresh
repo-and-packages docs and event-bus allow-list for setup-flow dialog splits.
