+++ app/lib/features/game/widgets/province_overlay/province_sea_zone_detail_overlay.dart
@@ -47,6 +47,8 @@ part 'province_sea_zone_detail_overlay_chrome.dart';
 part 'province_sea_zone_detail_overlay_tile_section_labels.dart';
 part 'province_sea_zone_detail_overlay_tile_section.dart';
 part 'province_sea_zone_detail_overlay_province_content.dart';
+part 'province_sea_zone_detail_overlay_province_content_unrevealed.dart';
+part 'province_sea_zone_detail_overlay_province_content_intel.dart';
 part 'province_sea_zone_detail_overlay_sea_zone_content.dart';
 part 'province_sea_zone_detail_overlay_economic_section.dart';
 part 'province_sea_zone_detail_overlay_military_section.dart';
