+++ tool/check_app_event_bus_decoupling.dart
@@ -108,7 +108,8 @@ const Set<String> _allowedFeatureLocalDialogFiles = <String>{
   // parent panel — `SPEC/program/app-ui-wiring.md` § "Local by design"
   // (`ResearchFundingBreakdownDialog`, Refs #3512).
   'app/lib/features/game/widgets/technology/research_slot_turn_preview_view_breakdown.dart',
-  'app/lib/features/shell/new_game_setup_flow_dialogs.dart',
+  'app/lib/features/shell/new_game_setup_flow_dialogs_error.dart',
+  'app/lib/features/shell/new_game_setup_flow_dialogs_progress.dart',
   // Split / move fleet — `SPEC/program/app-ui-wiring.md` "Split fleet" /
   // "Move fleet" paragraphs.
   'app/lib/features/game/widgets/units/naval/naval_units_panel.dart',
